requirements.txt
fastapi>=0.111,<1
uvicorn[standard]>=0.30,<1
slowapi>=0.1.9,<1
bcrypt>=4.1,<5
ldap3>=2.9,<3
pycryptodome>=3.20,<4
pyyaml>=6.0,<7
